Make_track is the rarest pepe
Wingat en hel del, mycket shared_ptr magi som jag inte riktigt förstår

+++ SuperFoobar/Block-filer/Non_breakable.cpp
@@ -1,5 +1,5 @@
 #include "Non_breakable.h"
 
-Non_breakable::Non_breakable(float x, float y, float height, float width)
+Non_breakable::Non_breakable(int x, int y, int height, int width)
 	:Block(x, y, height, width)
 {}
