+++ SuperFoobar/Character filer/Character.cpp
@@ -1,41 +1,41 @@
 #include "Character.h"
 
-float Character::get_x_pos() 
+int Character::get_x_pos()
 {
 	return x_pos;
 }
 
-float Character::get_y_pos() 
+int Character::get_y_pos()
 {
 	return y_pos;
 }
 
-void Character::move_x(float x_speed) 
+void Character::move_x(int x_speed)
 {
 	set_x_pos(x_speed);
 }
 
-void Character::move_y(float y_speed)
+void Character::move_y(int y_speed)
 {
 	set_y_pos(y_speed);
 }
 
-void Character::set_x_pos(float new_pos)
+void Character::set_x_pos(int new_pos)
 {
 	x_pos += new_pos;
 }
 
-void Character::set_y_pos(float new_pos)
+void Character::set_y_pos(int new_pos)
 {
 	y_pos += new_pos;
 }
 
-void Character::set_width(float new_width)
+void Character::set_width(int new_width)
 {
 	width = new_width;
 }
 
-void Character::set_height(float new_height)
+void Character::set_height(int new_height)
 {
 	height = new_height;
 }
