+++ SuperFoobar/Character filer/Enemy/Enemy_2.h
@@ -4,7 +4,7 @@ class Enemy_2 :
 	public Character
 {
 public:
-	Enemy_2(float x, float y, float width, float height)
+	Enemy_2(int x, int y, int width, int height)
 		: Character(x, y, width, height) {}
 	~Enemy_2();
 };
