+++ SuperFoobar/Interactable-filer/Power_up.cpp
@@ -2,6 +2,6 @@
 
 using namespace std;
 
-Power_up::Power_up(float x_pos, float y_pos, float height, float width)
+Power_up::Power_up(int x_pos, int y_pos, int height, int width)
 	: Interactable(x_pos, y_pos, height, width)
 {}
